+++ admin_app/src/components/TypingAnimation.tsx
